Chore: Arrangement domains #12
