+++ src/main/java/com/daigo/springboot_handson_4/domains/geocoder/ContentsGeoCoder.java
@@ -1,18 +1,19 @@
-package com.daigo.springboot_handson_4.cafedomains;
+package com.daigo.springboot_handson_4.domains.geocoder;
 
+import com.daigo.springboot_handson_4.domains.yahooapi.ResultInfo;
 import com.fasterxml.jackson.annotation.JsonProperty;
+import java.util.List;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-import java.util.List;
-
-@Setter
-@Getter
 @NoArgsConstructor
-public class LocalSearch {
+@Getter
+@Setter
+public class ContentsGeoCoder {
     @JsonProperty("ResultInfo")
     private ResultInfo resultInfo;
     @JsonProperty("Feature")
     private List<Feature> featureList;
 }
+
