+++ src/main/java/com/daigo/springboot_handson_4/domains/geocoder/Element.java
@@ -1,16 +1,16 @@
-package com.daigo.springboot_handson_4.cafedomains;
+package com.daigo.springboot_handson_4.domains.geocoder;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+@NoArgsConstructor
 @Setter
 @Getter
-@NoArgsConstructor
-public class Country {
-    @JsonProperty("Code")
-    private String code;
+public class Element {
     @JsonProperty("Name")
     private String name;
+    @JsonProperty("Kana")
+    private String kana;
 }
