+++ src/main/java/com/daigo/springboot_handson_4/domains/geocoder/Geometry.java
@@ -1,4 +1,4 @@
-package com.daigo.springboot_handson_4.domains;
+package com.daigo.springboot_handson_4.domains.geocoder;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Getter;
