+++ src/main/java/com/daigo/springboot_handson_4/domains/geocoder/Property.java
@@ -1,11 +1,10 @@
-package com.daigo.springboot_handson_4.domains;
+package com.daigo.springboot_handson_4.domains.geocoder;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
+import java.util.List;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
-
-import java.util.List;
 
 @NoArgsConstructor
 @Getter
