+++ src/main/java/com/daigo/springboot_handson_4/domains/localsearch/Country.java
@@ -1,4 +1,4 @@
-package com.daigo.springboot_handson_4.cafedomains;
+package com.daigo.springboot_handson_4.domains.localsearch;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Getter;
@@ -8,7 +8,7 @@ import lombok.Setter;
 @Setter
 @Getter
 @NoArgsConstructor
-public class Genre {
+public class Country {
     @JsonProperty("Code")
     private String code;
     @JsonProperty("Name")
