+++ src/main/java/com/daigo/springboot_handson_4/domains/localsearch/Genre.java
@@ -1,16 +1,16 @@
-package com.daigo.springboot_handson_4.domains;
+package com.daigo.springboot_handson_4.domains.localsearch;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-@NoArgsConstructor
 @Setter
 @Getter
-public class Element {
+@NoArgsConstructor
+public class Genre {
+    @JsonProperty("Code")
+    private String code;
     @JsonProperty("Name")
     private String name;
-    @JsonProperty("Kana")
-    private String kana;
 }
