+++ src/main/java/com/daigo/springboot_handson_4/domains/localsearch/Station.java
@@ -1,6 +1,6 @@
-package com.daigo.springboot_handson_4.cafedomains;
+package com.daigo.springboot_handson_4.domains.localsearch;
 
-import com.daigo.springboot_handson_4.domains.Geometry;
+import com.daigo.springboot_handson_4.domains.geocoder.Geometry;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
